requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
tensorflow==2.16.2
pillow==10.4.0
python-multipart==0.0.9
numpy==1.26.4
streamlit==1.38.0
requests==2.32.3
